Use dns hostname to reach the database
Docker rewrites /etc/hosts with IP/name mappings that allow containers
on the same network to refer to each other via names.

For local development, it's useful to just add the same mapping:

```
cat /etc/hosts
127.0.0.1       arangodb
```

With that in place the container will always reach the database via it's
name, whether it's dev or docker/prod.