Test db functions in an ephemeral database
We're adding a test for the database function(s) here but we've gone to
a little trouble to create a database, run the test(s) within it, and
then delete it after. This allows us to take advantage of Jest's ability
to run tests in parallel.

The modification to the Jest "testEnvironment" setting is because of
arangodb/arangojs#410.